feat: add leave requests and onboarding data structures with initial data
