+++ .gitignore
@@ -35,7 +35,10 @@ yarn-error.log*
 
 # vercel
 .vercel
+.claude
 
 # typescript
 *.tsbuildinfo
 next-env.d.ts
+
+
